refactor : 인물검색 통합
